requirements_test.txt
pytest==7.1.2
pytest-django==4.5.2
